Move CrossHostTransferBuffers into CommonPjRtClient
